+++ @ornikar/eslint-config-typescript-react/rules/react.js
@@ -1,5 +1,8 @@
 'use strict';
 
+// eslint-disable-next-line import/no-extraneous-dependencies
+const { rules: baseStyleRules } = require('eslint-config-airbnb-base/rules/style');
+
 module.exports = {
   rules: {
     'react/jsx-filename-extension': [
@@ -16,5 +19,8 @@ module.exports = {
     'react/prop-types': 'off',
     // https://github.com/yannickcr/eslint-plugin-react/issues/2702
     'react/sort-prop-types': 'off',
+
+    // overrides camelcase rule to allow UNSAFE_* commonly used in react libraries
+    camelcase: ['error', { ...baseStyleRules.camelcase[1], allow: ['^UNSAFE_'] }],
   },
 };