+++ @ornikar/eslint-config/tests/node/import-exports.mjs
@@ -1,0 +1,7 @@
+/* eslint-disable import/no-duplicates */
+
+// eslint-disable-next-line sort-imports
+import { b, a } from './exports.mjs';
+import { c, d } from './exports.mjs';
+
+console.log({ a, b, c, d });